seperated menuitems according to the menus

+++ db/migrate/20200524115432_add_menu_id_to_menuitems.rb
@@ -1,5 +1,5 @@
 class AddMenuIdToMenuitems < ActiveRecord::Migration[6.0]
   def change
-    add_column :menuitems, :menu_id, :integer
+    add_column :menuitems, :menu_id, :bigint
   end
 end
